cf978d solution

+++ c/cf/cf978d.cc
@@ -20,9 +20,9 @@ int main(int argc, char **argv) {
 			copy(bb, bb+n, aa);
 			aa[0] += ti[0];
 			aa[1] += ti[1];
-			int delta = aa[1]-aa[0];
+			int64_t delta = aa[1]-aa[0];
 			for (int i=2; i<n; i++) {
-				int d = aa[i]-(aa[i-1]+delta);
+				int64_t d = aa[i]-(aa[i-1]+delta);
 				if (d*d < 2) {
 					aa[i] -= d;
 					cnt += d*d;
